requirements.txt
opencv-python
numpy
pdf2image
Pillow
moondream